better trap handling.  this copies in most of pxa's trap.c's trap() (including fp instruction handling, though untested).
since there are three irq registers, trap handling is a bit different.
we always enable the bridge interrupt handler (e.g. the timer is
on the bridge), and make it call all the handlers (in the same way
as the bridge interrupt handler itself was called).

initial diff for proper trap() from salva.