cesa support for md5 & sha1.
this polls for completion.  that will probably be the way to go for
small packets.  but for larger ones, we will likely use the security
accelerator with dma.  will have to find out where the optimal point
to switch is.  i've done a few tests, but with datachache off:
262144 blocks md5/sha1 in 2450 ms (which is 1069 blocks (8552 bytes)
per clock tick).