add des support (polling only).
the caveat:  standard libinterp/keyring.c doesn't use the
des^(ECB CBC)^(encrypt decrypt) functions from libsec.  instead,
it uses block_cipher.  we only define the des* functions, so inferno
won't use this support with the default libinterp/keyring.c.  when
we want to use the accelerator with dma (and schedule the caller
away) we might have to be more careful, as not all callers might
expect that.

just as with the other algorithms, using the accelerator with tdma
comes later (if it's feasible).

triple des support is untested.  des ecb should pad short blocks
with zeros, but for now it assumes this won't happen.

this also renames some register defines to make a clearer distinction
between the various engines.  it also prevents an error() with lock held.