add a list of todo's.  disable des acceleration for now.
des is disabled since we don't derive a key schedule, but some
inferno functions use the expanded key directly by calling a function
we don't emulate.  in the future, those cases either have to be
changed, or we have to derive the key even when cesa.c doesn't use
it.