greatly improve buffer handling for ethernet, more similar to pc/ether82543gc.c.
we no longer allocate on transmit & receive.  at init of the driver
we allocate buffers for the entire rx ring.  at init we setup the
rx ring with blocks.  we set the free() function of the block to
our free function, so we can reuse those blocks (and never have to
allocate after initialisation).  for tx, we set up the descriptors
at startup, but no buffers.  on transmit of a block, we put the
block buffer in the next free descriptor, keep track of the block,
and kick the ethernet.  on interrupt we freeb the blocks that have
been sent.

this also removes the "tx ring full" errors i saw before.  operation
is now stable for both small & large ping floods.

the reason the code had been more complicated before this commit
is that i thought buffers in the rings had to be <= 1024 bytes.  i
had misread the docs.