some cleaning & little changes i had.
in the mkfile, do not depend on an awk that has strtonum.  openbsd's
(and plan 9's probably too) does not have it.  if you want to set
the KTZERO/KLOAD from command-line, just specify both.

in kernel config sheeva, keep some more in the root file system.
sdio is not reliable enough to rely on.  plus some cleaning.

mostly fix devsdio.c to not panic when booting without card that
works for the current code.  sdioinit() was calling sdinit(), which
fails with error() which is not allowed.  also add some small other
changes i had here.