first code that has read, erased and programmed the nand flash.
this is in a new file because it does most of what os/port/flashnand.c
does.  it seems os/port/flashnand.c executes older-style nand
commands (when 256 or 512 byte pages were normal, they are often
2048 bytes now and eg need more address cycles).

see the command on top of kwnand.c for more information.

this also disables network and sdio testing in misc/init, that is
for testing.  and add the flash partitions, see /dev/flash/* after
boot.  note that the uboot and kernel parts cannot be read because
they have a different style of ecc.