a struct for the l2 cache windows.  a command to half cpu frequency.  some more symbolic names for register bits that power down parts of the chip.
the l2 cache window registers can be used for setting uncachable
memory regions.  this isn't done yet.

halfing the frequency shows when running a cpu intensive task.
power stays about 0.5 watt lower.

unused parts are not yet powered down (aside from the first sata
port) because it does not seem to help much and it could be annoying
if someone is trying to use them in the future.