turn on mmu to get dcache.
the dcache does nothing if the mmu is not also enabled.  we don't
need the mmu for anything else, so we now just set up a direct
mapping between virtual and physical addresses (virtual modified
(vma) is equal to virtual & physical as well).  the first 512mb is
marked cacheable & bufferable (sdram), then only the register file
and nand region is set up, as non-cacheable, non-bufferable.  this
is done with "sections" of 1mb in the mmu.

the icache is turned on early, booting is now fast.
the l2 cache is turned off entirely, there are no public docs for it...

the drivers that do dma now have dcwb*, dcinv* and dcwbinv* calls
around the dma accesses.

due to faster operations, the mii/smi timeout had to be increased.

also prevent overflow during calculation of available memory.
increasing the heap & main to 256mb each (from 10mb)...