clean up devsdio.c some more.
the code should be more readablel, or closer match the sd specs.
for one, we now put the 136bit and 48bit responses in uvlongs,
keeping the (sometimes confusing) bit numbering (bits 7-1 are crc
and sometimes skipped in counting, sometimes not.  the uvlongs
always have them).

when executing an sd command, make the caller specify the expected
response type and the parameters for eg dma).  gets rid of functions
that try to guess, the caller should know.

some code for the scr register is added too, but doesn't yet work.
this register indicates whether the card understands 4bit communication.
so far, we've assumed it does.  haven't been able to read that
register from my sd card though...