introduce devbs.c, hot-pluggable block storage.  change devsdio.c to kwsdio.c and use devbs.c.
devsd.c is not used because its use of scsi stuff is annoying:  sd
cards and sata drives are not scsi.

this is not complete, but partitions for the sd card are set up and
one can read from them.  devsata.c will change to this at some point
as well.