handle Store.ready better.
we now list the disk directory even when disk has not been initialised.
we do allow rctl() and wctl() on such devices (e.g. to set "debug"
on the sdio driver before we try to init the disk).  we also make
sure to not lose memory/corrupt state when initialisation fails or
is done again.

if initing partitions fails, the data file can still be accessed.
eg to write new mbr with partition table to the data file.  a new
ctl "partinit" has been added to reinit the partition table without
having to reinit the device.

handle reading of non-aligned count and/or offset.  no writes,
perhaps they will be needed later.

actually write to disk when requested (instead of reading...).

also fix crash when closing a file that was open while the disk
changed underneath.

kwsdio.c:
- add debug wctl
- improve locking
- use anonymous static struct with lock and state
- some cosmetic changes.
- add wctl option "preerase", whether to erase before write (off by default).

part.c:
- allow both partition "plan9" and its child partitions.
- fix parsing table.  entries start with literal word "part".