make ncq work properly.
we now do proper tag accounting, and locking to get at the tags.

the trick is "uchar tags[32]", which is filled with tags.  when
someone wants to do io (sataread or satawrite), they take a qlock,
sleep until a tag is available, and use that tag for the request,
sleeping again until the command for that tag is finished (interrupt
handler does the wakeup).  a slot to send the ata command to the
disk is always available when a tag is available.

this has been tested by doing many (eg 1024) concurrent satareads.